requirements.txt
flask==2.3.3
pandas==2.0.3
plotly==5.16.1
python-dateutil==2.8.2
geopandas==0.13.2
shapely>=2.0.1,<3.0.0
